requirements.txt
pandas>=2.0,<2.3
numpy>=1.24
pyyaml>=6.0

scikit-learn>=1.4,<1.7
hdbscan>=0.8.33
tqdm>=4.66

transformers>=4.41,<4.45
sentence-transformers>=2.6,<3.1
torch>=2.1
